changed the ball postioin

+++ backend/src/blocked/block.service.ts
@@ -13,11 +13,7 @@ import { UserService } from "src/user/user.service";
 
 @Injectable()
 export class BlockService extends AbstractService {
-  constructor(
-    private friendService: FriendService,
-    private userService: UserService,
-    @InjectRepository(Block) private readonly blockRepository: Repository<Block>,
-  ) {
+  constructor(private friendService: FriendService, private userService: UserService, @InjectRepository(Block) private readonly blockRepository: Repository<Block>) {
     super(blockRepository);
   }
 
@@ -29,36 +25,34 @@ export class BlockService extends AbstractService {
     return block;
   }
 
-    async getBlockByUserids(senderId: number, receiverId: number) {
-	console.log(senderId, receiverId)
-      const sender = await this.userService.getUserById(senderId); 
-      const receiver = await this.userService.getUserById(receiverId); 
-      const block = await this.findOne(
-        {
-          sender: sender,
-          receiver: receiver
-        });
-      if (!block) {
-        throw new BadRequestException("This block doen't exist");
-      }
-      return block;
+  async getBlockByUserids(senderId: number, receiverId: number) {
+    console.log(senderId, receiverId);
+    const sender = await this.userService.getUserById(senderId);
+    const receiver = await this.userService.getUserById(receiverId);
+    const block = await this.findOne({
+      sender: sender,
+      receiver: receiver,
+    });
+    if (!block) {
+      throw new BadRequestException("This block doen't exist");
     }
+    return block;
+  }
 
-	async getBlocksFromUser(user: User) {
-		return await this.blockRepository.find({
-			where: {sender: user},
-			relations: ["receiver"]
-		});
-	}
+  async getBlocksFromUser(user: User) {
+    return await this.blockRepository.find({
+      where: { sender: user },
+      relations: ["receiver"],
+    });
+  }
 
   async getBlockBySenderAndReceiver(sender: User, receiver: User) {
     return await this.findOne({ sender: sender, receiver: receiver });
   }
 
-	async block(sender: User, receiver: User) {
-		const friendship = await this.friendService.getFriendshipByUserids(sender.id, receiver.id)
-		if (friendship)
-			await this.friendService.deleteFriendship(friendship);
-		return await this.create({sender: sender, receiver: receiver});
-	}
+  async block(sender: User, receiver: User) {
+    const friendship = await this.friendService.getFriendshipByUserids(sender.id, receiver.id);
+    if (friendship) await this.friendService.deleteFriendship(friendship);
+    return await this.create({ sender: sender, receiver: receiver });
+  }
 }
